requirements.txt
fastapi
numpy
uvicorn
image
pydantic
tensorflow-cpu
tensorflow_decision_forests
python-multipart
google-cloud-storage

